Fix mode symbol lookup using info-lookup-symbol and help-goto-info.
The mode was not previously providing a mapping between mode symbols
and their location in the Info manual.  This change adds this mapping
allowing info-lookup-symbol and help-goto-info to work with the
symbols documented in the manual.

Additionally, the formatting of user options in the manual has been
changed to be consistent with how they are documented in the Emacs
manual.