Fix defun and highlight for operator functions.
The defun name was not correctly displayed for operator functions.
Additionally, the operator function (string) name was not being
highlighted as a function name, but instead as a string.  Furthermore,
function and procedure calls were being highlighted using the
"function name" face instead of the "function call" face, which has
now been corrected.

Along with these corrections, font lock regression tests are being
introduced.  This uses the `ert-font-lock` package which is available
in Emacs 30 and later.  The font lock tests will be skipped if this
package is not available.

The Github Action has been modified to include more Emacs releases.
This was primarily done to include the "snapshot" release for Emacs 30
so that font lock testing can be performed.