Fix some function call highlighting issues.
This fixes a number of incorrect highlights surrounding function
calls.  Due to the ambiguity in the syntax tree, the function call
highlighting is not perfect, however a couple known highlighting
issues were fixed, including generic package instantiation, array
assignments, attribute function calls, etc.  Additionally, a
troubleshooting section was added for this topic to the documantation
to document the known issue and suggest workarounds.