Use major-mode-remap-defaults when available.
This was introduced in Emacs 30 and is preferred for the mode to use
this setting rather than updating major-mode-remap-alist, which is
more for use by the user.  The function major-mode-remap was
introduced at the same time and we use the existence of this function
to know when to update which list.