Update font lock rules based on recent allocator grammar correction.
An issue was corrected in the Ada grammar surrounding allocators (so
discriminant constraints are no longer parsed as function calls).  As
part of that change, the subtype_mark field name was added within the
subtype indication, therefore it's no longer necessary to have a
special match rule for finding the subtype name in an allocator
expression.  That match rule also was not perfect as it didn't account
for a possible leading null exclusion or subpool specification.
Therefore, using the subtype_mark field name is much more accurate.

Font lock tests have been added for checking the highlighting in many
different allocator expressions.