Add support for LSP workspaces with external directories (#5).
Utilize the Ada Language Server's "als-source-dirs" command to query
the set of source directories in the project file and configure
lsp-mode to treat source directories outside the workspace as "library
folders".  This allows the source files in those directories to be
opened using the same Language Server instance as the base project,
rather than spawning a new Language Server instance for the external
directories, which is necessary for navigation, hover information,
etc. to work correctly.