Handle aggressive indent of comments specially.
For aggressive indenting, adjacent comments will be indented as a
group, rather than causing the parent of the comment to be indented.
Indenting the parent can be expensive for large packages, subprograms,
etc., and it is likely more intuitive that a comment be treated the
equivalent of a statement for indenting purposes.