Fix numerous best-effort indentation issues.
Ada keywords which are out-dented have been electrified (e.g.,
`begin`, `end`, etc).  When `electric-indent-mode` is enabled but
`RET` is remapped (such as to `newline-and-indent` in Doom Emacs),
it's not guaranteed that the current line will be re-indented when
pressing `RET`.  As a result, electric indentation is necessary to
cause out-denting.

Numerous tests were added to verify the electrified keyword behavior,
uncovering a number of issues in the best-effort indentation which
were addressed.