Make grammar uniformly configurable.
Some Emacs distributions (e.g., Doom Emacs) expect to be able to
configure all tree-sitter grammars through the use of the standard
`treesit-language-source-alist` rather than a mode-specific
customization.  Therefore, `treesit-language-source-alist` is no
longer initialized buffer locally during mode initialization, and
instead the global version is configured for Ada, based on
`ada-ts-mode-grammar`, when the package is loaded.

The grammar specified in `ada-ts-mode-grammar` is appended to the end
of `treesit-language-source-alist` so that any user configuration that
may exist closer to the front of the list will take priority.