Prefer textDocument/formatting for whole buffer LSP indentation.
When the LSP back-end is being used for indentation, and the range
selected for indentation is the entire buffer, use the LSP request
`textDocument/formatting` instead of the alternate range-based request
`textDocument/rangeFormatting`.  It seems range formatting doesn't
work as well with the Ada Language Server, especially near the the top
and bottom of the document, so preferring `textDocument/formatting`
results in better indentation.